+++ src/process_utility.c
@@ -232,6 +232,7 @@ check_alter_table_allowed_on_ht_with_compression(Hypertable *ht, AlterTableStmt 
 			case AT_SetStatistics: /* should this be pushed down in some way? */
 			case AT_AddColumn:	 /* this is passed down */
 			case AT_ColumnDefault: /* this is passed down */
+			case AT_DropColumn:	/* this is passed down */
 #if PG14_GE
 			case AT_ReAddStatistics:
 			case AT_SetCompression: