+++ CHANGELOG.md
@@ -19,6 +19,7 @@ set `timescaledb.pushdown_functions = 'whitelisted'` on the access node.
 
 **Features**
 * #3768 Allow ALTER TABLE ADD COLUMN with DEFAULT on compressed hypertable
+* #3769 Allow ALTER TABLE DROP COLUMN on compressed hypertable
 
 **Bugfixes**
 * #3766 Fix segfault in ts_hist_sfunc